refactor(collapse): extract useBelowBreakpoint from Rail and export it
Moves the viewport-breakpoint hook out of Rail.tsx into the shared
_internal/collapse module and re-exports it from the package root, so
AppLayout's later overlay-sidebar mode can consume the same hook instead
of duplicating it. Also extracts the matchMedia jsdom stub shared between
collapse.test.ts and Rail.test.tsx into _internal/matchMediaStub.ts.
